Working mqtt and various fixes

+++ Config.h
@@ -23,15 +23,15 @@ uint8_t SmaBTAddress[6]= {0x00, 0x80, 0x25, 0x27, 0x38, 0xE7}; // my SMA SMC6000
 
 
 #include <WiFi.h>
-#include <WiFiClient.h>
 #include <WebServer.h>
 #include <DNSServer.h>
 #include <ESPmDNS.h>
 #include <Preferences.h>
+#include <PubSubClient.h>
 
 struct Config {
   String mqttBroker;
-  int mqttPort;
+  uint16_t mqttPort;
   String mqttUser;
   String mqttPasswd;
   String mqttTopic;
